+++ css/style.css
@@ -1061,6 +1061,26 @@ fieldset {
 *Сообщения об ошибке загрузки данных
 */
 .data__error {
-  color: darkred;
+  position: relative;
+  top: 100%;
+  left: 100%;
+  z-index: 1100;
+
+  width: 100%;
+  height: 20%;
+
+  padding-bottom: 10px;
+
   text-align: center;
+  vertical-align: middle;
+
+  background-color: rgba(0, 0, 0, 0.8);
+  color: white;
+  text-align: center;
+}
+
+.data__error__button {
+  color: darkred;
+  font-size: 14px;
+  text-decoration: underline;
 }
