+++ index.html
@@ -237,8 +237,8 @@
   <!-- Сообщение об ошибке загрузки данных -->
   <template id="data-error">
     <div class="data__error">
-      <p class="data__error__message">Не удалось загрузить похожие обьявления :( </p>
-
+      <p class="data__error__message">Не удалось загрузить похожие обьявления :( <br></p>
+      <button id="data-error-retry-button" type="button" class="data__error__button">Попробовать еще раз</button> или <button id="data-error-close-button" class="data__error__button">убрать сообщение</button>
     </div>
   </template>
   <script src="pristine/pristine.min.js"></script>
