+++ js/form-validation.js
@@ -1,5 +1,5 @@
 import { sendData } from './api.js';
-import { mapReset } from './map.js';
+import { mapReset } from './map-util.js';
 import { resetSlider } from './slider.js';
 import { blockSubmitButton, formSuccess, formError } from './form-util.js';
